chore: migrate to Express v5 in the tests
This moves us to use Express v5 in the tests. Only OpenTelemetry needed
a little more work and I believe this is because async routing is now
properly supported in Express. It means that we can no longer shut down
the OpenTelemetry SDK _before_ the response has been sent, because the
trace is sent on response close rather than once the middleware finishes
executing.