feat: allow config of HTTP server duration buckets
This allows users to customise the buckets in the HTTP server duration
histogram. This will only come into play when the option is explicitly
set (via JS configurations or environment variable), this means we can
safely make this change and the behaviour across all existing projects
will not change.

This resolves #1486.