feat: add more e2e tests
those are showing very specifically what we are testing (the body of
the response and invalid JSON and long body) instead of having the tests
just being part of the other tests. We can still keep checking the body
of the response as part of the pre existing tests because the body
is now set with the new code and it does not create too much noise.
But having new tests for the new situations make it easier to understand
what this PR is adding.