feat: replace custom cause with native cause
This is possible now Node 18+ is required.
The native cause are being passed as unknown in the Error class so we
are keeping that behaviour, and accepting anything for a cause.
Our custom cause was setting it as null when nothing was passed. This
is no longer the case and if there is no cause, it will just be absent
from the error.