feat: update serializer with non error cause
We have previously updated our BaseError (etc.) to now use the cause
provided by standard error. The cause is type unknown, and does not have
to be of type Error. That means we need our serializer to accept those
cause. We also update types to not have to set cause as null and to
get the type type as Error cause.